KP-5973 Do unattended configuration with authorization token
Trying to configure a runner without `--unttended` fails in pod due to
the script determining that it cannot write to terminal. For some reason
the unattended configuration disables the option to use a personal
access token though, so we must fetch an authorization token from the
GitHub API first and then use it when creating the runner.