KP-5973 Create separate user for the runner
When running inside a pod on Rahti, we do not have root access. Thus we
need to extract the runner so that the resulting files are readable and
executable by a user with UID greater than 1031660000. We still need to
do some things (e.g. installations using yum) as root, but the rest are
now done with a dedicated runner user.