KP-6208 Create Ubuntu-based runners
This is required because on CentOS 7 it's hard to get new-enough Python to fulfill
Airflow needs. Thus we are now testing and intend to run on production
using Ubuntu instead.

The Ubuntu 22.04 Docker image apparently does not come with Python
preinstalled, so it is installed hand. For easy using, it is also
symlinked to be invoked via `python` in addition to `python3`.