KP-6208 Install Python 3.8
This is the newest version that Airflow supports. No
`update-alternatives` or similar operations were done; this way the same
runner can be adopted for use in other repositories too when we get
support for that. On the other hand, this means that invoking Python 3.8
must be done explicitly using `python3.8` instead of just `python3` or
`python`. Similarly, to install software for a specific version using
pip, one must call pip as a module: `python3.8 -m pip install ...`.