KP-7378 Prepare for having different runners
This is useful because it's good to have the Python unit test runner be
reasonably vanilla: that way it's less likely that preinstalled packages
mask missing requirements definitions or cause us to use inconsistent
versions of packages. This makes it more likely that the automatic test
runs match the production environment.

Renamed the relevant directories to be plural, made the file names more
specific and changed the READMEs accordingly.